fix(cuelite): bound-check list index and MinRunes in wide-integer space
CodeQL flagged int64-to-int conversions before the bounds checks in
list indexing and strings.MinRunes. On 32-bit targets (the plan-240
wasm build) a wide literal would truncate and could select a valid
wrong element or invert the rune-count check. Compare in int64/float64
space before any narrowing.

+++ cue/cuelite/unify.go
@@ -172,7 +172,9 @@ func concreteSatisfiesBound(c *engineValue, bd bound) bool {
 	case opNotMatch:
 		return c.kind == kString && !bd.re.MatchString(c.str)
 	case opMinRunes:
-		return c.kind == kString && utf8.RuneCountInString(c.str) >= int(bd.num)
+		// Compare in float64 space: int(bd.num) would truncate a huge
+		// MinRunes argument on 32-bit targets (wasm) and invert the check.
+		return c.kind == kString && float64(utf8.RuneCountInString(c.str)) >= bd.num
 	}
 	if bd.isStr {
 		if c.kind != kString {
